added branchesGrid component
